requirements.txt
airtest==1.2.7
APScheduler==3.10.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==22.1.0
backports.zoneinfo==0.2.1
bcrypt==4.0.1
BeautifulReport==0.1.3
beautifulsoup4==4.11.1
Brotli==1.0.9
cached-property==1.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
ConfigArgParse==1.5.3
cryptography==38.0.4
decorator==5.1.1
Deprecated==1.2.13
Django==4.1.3
django-apscheduler==0.6.2
django-mysql==4.7.1
django-redis==5.2.0
django-simpleui==2022.12.18
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docker==4.4.4
docker-compose==1.25.4
dockerpty==0.4.1
docopt==0.6.2
drf-extensions==0.7.1
drf-spectacular==0.24.2
dubboclient==1.0.1
exceptiongroup==1.0.4
facebook-wda==1.4.6
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
gevent==22.10.2
geventhttpclient==2.0.8
gitdb2==2.0.6
GitPython==2.1.15
greenlet==2.0.1
hrpc==1.0.9
idna==3.4
importlib-metadata==6.0.0
inflection==0.5.1
iniconfig==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==3.2.0
locust==2.14.2
MarkupSafe==2.1.1
msgpack==1.0.4
mss==6.1.0
mysqlclient==2.1.1
nose==1.3.7
numpy==1.23.5
opencv-contrib-python==4.6.0.66
opencv-python==4.6.0.66
packaging==21.3
paramiko==2.12.0
Pillow==9.3.0
pipreqs==0.4.11
pluggy==1.0.0
pocoui==1.0.88
psutil==5.9.4
py==1.11.0
pyaml==19.12.0
pycparser==2.21
PyGithub==1.45
PyJWT==2.6.0
PyMySQL==1.0.2
PyNaCl==1.5.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytest==7.2.0
python-dateutil==2.8.2
python-gitlab==1.15.0
python-hessian==1.1.2
python-jenkins-monitor==1.4.1
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
pywin32==227
pywinauto==0.6.3
PyYAML==5.3
pyzmq==25.0.0
redis==4.4.0
requests==2.28.1
retry==0.9.2
roundrobin==0.0.4
six==1.16.0
smmap==5.0.0
smmap2==3.0.1
soupsieve==2.3.2.post1
sqlparse==0.4.3
svn==1.0.1
texttable==1.6.7
tomli==2.0.1
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
websocket-client==0.48.0
Werkzeug==2.2.2
wfastcgi==3.0.0
wrapt==1.14.1
yarg==0.1.9
zipp==3.12.0
zope.event==4.6
zope.interface==5.5.2
